AIoT-Monitors/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
paramiko==3.3.1
SQLAlchemy==2.0.20
python-dotenv==1.0.0
pytest==7.4.2 